% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{filter.annotation.biseq}
\alias{filter.annotation.biseq}
\title{filter.annotation.biseq}
\usage{
filter.annotation.biseq(rnb.set, snp = TRUE, snp.list, somatic = TRUE,
  qual.filter = NULL)
}
\arguments{
\item{rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} containg annotation information.}

\item{snp}{Flag indicating if snps are to be removed. Either SNPs annotated in the RnBeads annotation object of specified as an
additional file with one SNP identifier per row in \code{snp.list}.}

\item{snp.list}{Path to a file containing known SNPs. One SNP identifier should be present per row.}

\item{somatic}{Flag indicating if only somatic probes are to be kept.}

\item{qual.filter}{Vector of indices removed during quality filtering.}
}
\value{
A vector of indices of sites surviving the annotation filter criteria.
}
\description{
This function removes sites in SNP or sex-chromosomal context.
}
\details{
This functions uses information on the sites on the chip and transfers this knowledge to BS data sets.
}
