% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{filter.nas}
\alias{filter.nas}
\title{filter.nas}
\usage{
filter.nas(rnb.set, subs, qf.qual)
}
\arguments{
\item{rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} containing methylation information.}

\item{subs}{Optional argument specifying the subset of samples to be used}

\item{qf.qual}{Vector of indices that survided quality filtering}
}
\value{
Vector of indices that survived NA filtering.
}
\description{
This function removes any site that contains a missing methylation value in a sample.
}
