% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{filter.quality.covg}
\alias{filter.quality.covg}
\title{filter.quality.covg}
\usage{
filter.quality.covg(rnb.set, min.covg, min.covg.quant = 0.05,
  max.covg.quant = 0.95)
}
\arguments{
\item{rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} containing the CpG sites used for filtering, all well as intensity and
coverage informatio, if provided.}

\item{min.covg}{Integer specifying the minimum number of beads required for including the site in the analysis.}

\item{min.covg.quant}{Lower quantile of coverage which is to be removed.}

\item{max.covg.quant}{Upper quantile of coverage which is to be removed.}
}
\value{
The indices of the sites that are to be kept.
}
\description{
This functions filters the CpG sites in the given rnb.set for quality criteria specified in the arguments.
}
\details{
The probes with lower/higher coverage than \code{min.int.quant}/
           \code{max.int.quant} are removed.
}
