% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{filter.quality}
\alias{filter.quality}
\title{filter.quality}
\usage{
filter.quality(
  rnb.set,
  beads = TRUE,
  min.beads,
  intensity = TRUE,
  min.int.quant = 0.1,
  max.int.quant = 0.95,
  subs
)
}
\arguments{
\item{rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} containing the CpG sites used for filtering, all well as intensity and
coverage informatio, if provided.}

\item{beads}{Flag indicating if sites not having more than \code{min.beads} number of beads in all samples are to be removed.}

\item{min.beads}{Integer specifying the minimum number of beads required for including the site in the analysis.}

\item{intensity}{Flag indicating if sites are to be filtered according to their intensity information.}

\item{min.int.quant}{Lower quantile of intensities which is to be removed.}

\item{max.int.quant}{Upper quantile of intensities which is to be removed.}

\item{subs}{Argument specifying the subset of samples to be used}
}
\value{
The indices of the sites that are to be kept.
}
\description{
This functions filters the CpG sites in the given rnb.set for quality criteria specified in the arguments.
}
\details{
If \code{intensity} is set, those probes with lower/higher intensity in one of the channels than \code{min.int.quant}/
           \code{max.int.quant} are removed.
}
