% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepare.data}
\alias{prepare.data}
\title{FUNCTIONS
prepare.data}
\usage{
prepare.data(
  rnb.set,
  work.dir = getwd(),
  analysis.name = "analysis",
  sample.selection.col = NA,
  sample.selection.grep = NA,
  pheno.columns = NA,
  id.column = rnb.getOption("identifiers.column"),
  normalization = "none",
  ref.ct.column = NA,
  ref.rnb.set = NULL,
  ref.rnb.ct.column = NA,
  prepare.true.proportions = FALSE,
  true.A.token = NA,
  houseman.A.token = NA,
  estimate.houseman.prop = FALSE,
  filter.beads = !is.null(rnb.set@covg.sites),
  min.n.beads = 3,
  filter.intensity = inherits(rnb.set, "RnBeadRawSet"),
  min.int.quant = 0.001,
  max.int.quant = 0.999,
  filter.na = TRUE,
  filter.context = TRUE,
  filter.snp = TRUE,
  filter.somatic = TRUE,
  filter.cross.reactive = T,
  remove.ICA = F,
  conf.fact.ICA = NULL,
  ica.setting = NULL,
  snp.list = NULL,
  execute.lump = FALSE,
  dist.snps = FALSE
)
}
\arguments{
\item{rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} for which analysis is to be performed.}

\item{work.dir}{A path to a existing directory, in which the results are to be stored}

\item{analysis.name}{A string representing the dataset for which analysis is to be performed. Only used to create a folder with a 
descriptive name of the analysis.}

\item{sample.selection.col}{A column name in the phenotypic table of \code{rnb.set} used to selected a subset of samples for
analysis that contain the string given in \code{sample.selection.col}.}

\item{sample.selection.grep}{A string used for selecting samples in the column \code{sample.selection.grep}.}

\item{pheno.columns}{Vector of column names in the phenotypic table of \code{rnb.set} that is kept and exported for further 
exploration.}

\item{id.column}{Sample-specific ID column name in \code{rnb.set}}

\item{normalization}{Normalization method to be performed before employing MeDeCom. Can be one of \code{"none", "dasen", "illumina", "noob", "bmiq"}.}

\item{ref.ct.column}{Column name in \code{rnb.set} used to extract methylation information on the reference cell types.}

\item{ref.rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} containing methylation information on reference cell types.}

\item{ref.rnb.ct.column}{Column name in \code{ref.rnb.set} used to extract methylation information on the reference cell types.}

\item{prepare.true.proportions}{Flag indicating if true proportions are either available in \code{rnb.set} or to be estimated 
with Houseman's reference-based deconvolution approach.}

\item{true.A.token}{String present in the column names of \code{rnb.set} used for selecting the true proportions of the corresponding
cell types.}

\item{houseman.A.token}{Similar to \code{true.A.token}, but not containing the true proportions, rather the estimated proportions
by Houseman's method.}

\item{estimate.houseman.prop}{If neither \code{true.A.token} nor \code{houseman.A.token} are given, the proportions of the reference
cell type are estimated with Houseman's approach.}

\item{filter.beads}{Flag indicating, if site-filtering based on the number of beads available is to be conducted.}

\item{min.n.beads}{Minimum number of beads required in each sample for the site to be considered for adding to MeDeCom.}

\item{filter.intensity}{Flag indicating if sites should be removed according to the signal intensities (the lowest and highest quantiles
given by \code{min.int.quant} and \code{max.int.quant}). Note that all sites are removed that have a value outside of
the provided quantile range in either of the channels and in any of the samples.}

\item{min.int.quant}{Lower quantile of intensities which is to be removed.}

\item{max.int.quant}{Upper quantile of intensities which is to be removed.}

\item{filter.na}{Flag indicating if sites with any missing values are to be removed or not.}

\item{filter.context}{Flag indicating if only CG probes are to be kept.}

\item{filter.snp}{Flag indicating if annotated SNPs are to be removed from the list of sites according to RnBeads' SNP list. Or as the sites
specified in \code{snp.list}.}

\item{filter.somatic}{Flag indicating if only somatic probes are to be kept.}

\item{filter.cross.reactive}{Flag indicating if sites showing cross reactivity on the array are to be removed.}

\item{remove.ICA}{Flag indicating if independent component analysis is to be executed to remove potential confounding factor.
If \code{TRUE},conf.fact.ICA needs to be specified.}

\item{conf.fact.ICA}{A vector of column names in the sample annotation sheet representing potential confounding factors.}

\item{ica.setting}{Named vector of settings passed to run.rnb.ica. Options are \code{nmin, nmax, thres.sd, alpha.fact, save.report, alpha.feat, type, ncores}. See 
\code{\link{run.rnb.ica}} for further details. NULL indicates the default setting.}

\item{snp.list}{Path to a file containing CpG IDs of known SNPs to be removed from the analysis, if \code{filter.snp} is \code{TRUE}.}

\item{execute.lump}{Flag indicating if the LUMP algorithm is to be used for estimating the amount of immune cells in a particular sample.}

\item{dist.snps}{Flag indicating if SNPs are to removed by determining if the pairwise differences between the CpGs in the samples are trimodally
distributed as it is frequently found around SNPs.}
}
\value{
A list with four elements: \itemize{
          \item quality.filter The indices of the sites that survived quality filtering
}
}
\description{
This functions prepares Illumina BeadChip data for a MeDeCom/EDec/RefFreeCellMix run.
}
\author{
Michael Scherer, Pavlo Lutsik
}
