% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpG_subsets.R
\name{prepare_CG_subsets}
\alias{prepare_CG_subsets}
\alias{prepare.CG.subsets}
\title{prepare.CG.subsets}
\usage{
prepare.CG.subsets(
  meth.data = NULL,
  rnb.set = NULL,
  marker.selection,
  n.markers = 5000,
  remove.correlated = FALSE,
  cor.threshold = "quantile",
  write.files = FALSE,
  out.dir = NA,
  ref.rnb.set = NULL,
  ref.pheno.column = NULL,
  n.prin.comp = 10,
  range.diff = 0.05,
  custom.marker.file = "",
  store.heatmaps = F,
  heatmap.sample.col = NULL,
  K.prior = NULL
)
}
\arguments{
\item{meth.data}{A \code{matrix} or \code{data.frame} containing methylation information. If NULL, methylation information needs to be provided
through \code{rnb.set}}

\item{rnb.set}{An object of type \code{\link{RnBSet-class}} containing methylation, sample and optional coverage information.}

\item{marker.selection}{A vector of strings representing marker selection methods. Available method are \itemize{
         \item{"\code{all}"} Using all sites available in the input.
         \item{"\code{pheno}"} Selected are the top \code{n.markers} site that differ between the phenotypic
                groups defined in data preparation or by \code{\link{rnb.sample.groups}}. Those are
                selected by employing limma on the methylation matrix.
         \item{"\code{houseman2012}"} 50k sites determined to be cell-type specific for blood cell types using the Houseman's reference-
                based deconvolution and the Reinius et al. reference data set. See Houseman et.al. 2012 and Reinis et.al. 2012.
                NOTE: This option should only be used for whole blood data generated using the 450k array.
         \item{"\code{houseman2014}"} Selects the sites said to be linked to cell type composition by \code{RefFreeEWAS},
                which is similar to surrogate variable analysis. See Houseman et.al. 2014.
         \item{"\code{jaffe2014}"} The 600 sites stated as related to cell-type composition Jaffe et.al. 2014.
                NOTE: This option should only be used for whole blood data generated using the 450k array.
         \item{"\code{rowFstat}"} Markers are selected as those found to be associated to the reference cell
                types with F-statistics. If this option is selected, \code{ref.rnb.set} and \code{ref.pheno.column}
                need to be specified.
         \item{"\code{random}"} Sites are randomly selected.
         \item{"\code{pca}"} Sites are selected as those with most influence on the principal components.
         \item{"\code{var}"} Selects the most variable sites.
         \item{"\code{hybrid}"} Selects (n.markers/2) most variable and (n.markers/2) random sites.
         \item{"\code{range}"} Selects the sites with the largest difference between minimum and maximum
              across samples.
         \item{"\code{pcadapt}"} Uses principal component analysis as implemented in the \code{"bigstats"}
              R package to determine sites that are significantly linked to the potential cell types. This
              requires specifying K a priori (argument \code{K.prior}). We thank Florian Prive and Sophie
              Achard for providing the idea and parts of the codes.
         \item{"\code{edec_stage0}} Employs EDec's stage 0 to infer cell-type specific markers. By default
              EDec's example reference data is provided. If a specific data set is to be provided, it needs
              to be done through \code{ref.rnb.set}.
         \item{"\code{custom}"} Specifying a custom file with indices.
}}

\item{n.markers}{The number of sites to be selected. Defaults to 5000.}

\item{remove.correlated}{Flag indicating if highly correlated sites are to be removed}

\item{cor.threshold}{Numeric indicating a threshold above which sites are not to be considered in the feature selection.
If \code{"quantile"}, sites correlated higher than the 95th quantile are removed.}

\item{write.files}{Flag indicating if the selected sites are to be stored on disk.}

\item{out.dir}{Path to the working directory used for analyis, or data preparation.}

\item{ref.rnb.set}{An object of type \code{\link{RnBSet-class}} or a path to such an object stored on disk, 
if \code{rowFstat} is selected.}

\item{ref.pheno.column}{Optional argument stating the column name of the phenotypic table of \code{ref.rnb.set} with
the reference cell type.}

\item{n.prin.comp}{Optional argument deteriming the number of prinicipal components used for selecting the most important sites.}

\item{range.diff}{Optional argument specifying the difference between maximum and minimum required.}

\item{custom.marker.file}{Optional argument containing an absolute path to a file that specifies the indices used for employing MeDeCom. Can be provided
either as an \code{RDS} file containing a vector of indices to select or as a \code{txt, csv, tsv} file containing each index
to be selected as a single row.}

\item{store.heatmaps}{Flag indicating if a heatmap of the selected input sites is to be create from the input methylation matrix.
The files are then stored in the 'heatmaps' folder in out.dir.}

\item{heatmap.sample.col}{Column name in the phenotypic table of \code{rnb.set}, used for creating a color scheme in the heatmap.}

\item{K.prior}{K determined from visual inspection. Only has an influence, if \code{marker.selection="pcadapt"}.}
}
\value{
List of indices, one entry for each marker selection method specified by \code{marker.selection}. The indices correspond
         to the sites that should be used in \code{rnb.set}.
}
\description{
This routine selects a subset of CpGs sites used for MeDeCom analysis. Different selection methods are supported.
}
\details{
For methods "\code{houseman2012}" and "\code{jaffe2014}", a predefined set of markers is used. Since those correspond to
         absolute indices on the chip, the provided \code{rnb.set} must not be preprocessed and therefore still contain all sites.
         For the other metods, you may used \code{\link{prepare.data}} to filter sites for quality and context.
}
\references{
\itemize{
            \item{1.} Houseman, E. A., Accomando, W. P., Koestler, D. C., Christensen, B. C., Marsit, C. J., Nelson, H. H., ..., Kelsey, K.
                T. (2012). DNA methylation arrays as surrogate measures of cell mixture distribution. BMC Bioinformatics, 13. 
            \item{2.} Reinius, L. E., Acevedo, N., Joerink, M., Pershagen, G., Dahlen, S. E., Greco, D., ..., A., & Kere, J.
                   (2012). Differential DNA methylation in purified human blood cells: Implications for cell lineage and studies on disease susceptibility.
                   PLoS ONE, 7(7). https://doi.org/10.1371/journal.pone.0041361
            \item{3.} Houseman, E. A., Molitor, J., & Marsit, C. J. (2014). Reference-free cell mixture adjustments in analysis of 
                DNA methylation data. Bioinformatics, 30(10), 1431-1439. https://doi.org/10.1093/bioinformatics/btu029
            \item{4.} Jaffe, A. E., & Irizarry, R. A. (2014). Accounting for cellular heterogeneity is critical in epigenome-wide 
                association studies. Genome Biology, 15(2), R31. https://doi.org/10.1186/gb-2014-15-2-r31
                 
}
}
\author{
Michael Scherer, Pavlo Lutsik
}
