% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepare_data}
\alias{prepare_data}
\title{GLOBALS
FUNCTIONS
prepare_data}
\usage{
prepare_data(RNB_SET, WORK_DIR, analysis.name, SAMPLE_SELECTION_COL = NA,
  SAMPLE_SELECTION_GREP = NA, PHENO_COLUMNS = NA,
  ID_COLUMN = rnb.getOption("identifiers.column"),
  NORMALIZATION = "none", REF_CT_COLUMN = NA, REF_RNB_SET = NA,
  REF_RNB_CT_COLUMN = NA, PREPARE_TRUE_PROPORTIONS = FALSE,
  TRUE_A_TOKEN = NA, HOUSEMAN_A_TOKEN = NA,
  ESTIMATE_HOUSEMAN_PROP = FALSE,
  FILTER_BEADS = !is.null(RNB_SET@covg.sites), MIN_N_BEADS = 3,
  FILTER_INTENSITY = inherits(RNB_SET, "RnBeadRawSet"),
  MIN_INT_QUANT = 0.1, MAX_INT_QUANT = 0.95, FILTER_NA = TRUE,
  FILTER_CONTEXT = TRUE, FILTER_SNP = TRUE, FILTER_SOMATIC = TRUE,
  snp.list = NULL)
}
\arguments{
\item{RNB_SET}{An object of type \code{\link[RnBeads]{RnBSet-class}} for which analysis is to be performed.}

\item{WORK_DIR}{A path to a existing directory, in which the results are to be stored}

\item{analysis.name}{A string representing the dataset for which analysis is to be performed. Only used to create a folder with a 
descriptive name of the analysis.}

\item{SAMPLE_SELECTION_COL}{A column name in the phenotypic table of \code{RNB_SET} used to selected a subset of samples for
analysis that contain the string given in \code{SAMPLE_SELECTION_GREP}.}

\item{SAMPLE_SELECTION_GREP}{A string used for selecting samples in the column \code{SAMPLE_SELECTION_COL}.}

\item{PHENO_COLUMNS}{Vector of column names in the phenotypic table of \code{RNB_SET} that is kept and exported for further 
exploration.}

\item{ID_COLUMN}{Sample-specific ID column name in \code{RNB_SET}}

\item{NORMALIZATION}{Normalization method to be performed before employing MeDeCom. Can be one of \code{"none","dasen","illumina","noob"}.}

\item{REF_CT_COLUMN}{Column name in \code{RNB_SET} used to extract methylation information on the reference cell types.}

\item{REF_RNB_SET}{An object of type \code{\link[RnBeads]{RnBSet-class}} containing methylation information on reference cell types.}

\item{REF_RNB_CT_COLUMN}{Column name in \code{REF_RNB_SET} used to extract methylation information on the reference cell types.}

\item{PREPARE_TRUE_PROPORTIONS}{Flag indicating if true proportions are either available in \code{RNB_SET} or to be estimated 
with Houseman's reference-based deconvolution approach.}

\item{TRUE_A_TOKEN}{String present in the column names of \code{RNB_SET} used for selecting the true proportions of the corresponding
cell types.}

\item{HOUSEMAN_A_TOKEN}{Similar to \code{TRUE_A_TOKEN}, but not containing the true proportions, rather the estimated proportions
by Houseman's method.}

\item{ESTIMATE_HOUSEMAN_PROP}{If neither \code{TRUE_A_TOKEN} nor \code{HOUSEMAN_A_TOKEN} are given, the proportions of the reference
cell type are estimated with Houseman's approach.}

\item{FILTER_BEADS}{Flag indicating, if site-filtering based on the number of beads available is to be conducted.}

\item{MIN_N_BEADS}{Minimum number of beads required in each sample for the site to be considered for adding to MeDeCom.}

\item{FILTER_INTENSITY}{Flag indicating if sites should be removed according to the signal intensities (the lowest and highest quantiles
given by \code{MIN_INT_QUANT} and \code{MAX_INT_QUANT}).}

\item{MIN_INT_QUANT}{Lower quantile of intensities which is to be removed.}

\item{MAX_INT_QUANT}{Upper quantile of intensities which is to be removed.}

\item{FILTER_NA}{Flag indicating if sites with any missing values are to be removed or not.}

\item{FILTER_CONTEXT}{Flag indicating if only CG probes are to be kept.}

\item{FILTER_SNP}{Flag indicating if annotated SNPs are to be removed from the list of sites according to RnBeads' SNP list. (@TODO: we
could provide an addititional list of SNPs, similar to RnBeads blacklist for filtering)}

\item{FILTER_SOMATIC}{Flag indicating if only somatic probes are to be kept.}

\item{snp.list}{Path to a file containing CpG IDs of known SNPs to be removed from the analysis, if \code{FILTER_SNP} is \code{TRUE}.}
}
\value{
A list with four elements: \itemize{
          \item quality.filter The indices of the sites that survived quality filtering
}
}
\description{
This functions prepares Illumina BeadChip data for a MeDeCom run.
}
