% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepare_data_BS}
\alias{prepare_data_BS}
\title{prepare_data_BS}
\usage{
prepare_data_BS(
  RNB_SET,
  WORK_DIR = getwd(),
  analysis.name = "analysis",
  SAMPLE_SELECTION_COL = NA,
  SAMPLE_SELECTION_GREP = NA,
  REF_CT_COLUMN = NA,
  PHENO_COLUMNS = NA,
  PREPARE_TRUE_PROPORTIONS = FALSE,
  TRUE_A_TOKEN = NA,
  HOUSEMAN_A_TOKEN = NA,
  ID_COLUMN = rnb.getOption("identifiers.column"),
  FILTER_COVERAGE = hasCovg(RNB_SET),
  MIN_COVERAGE = 5,
  MIN_COVG_QUANT = 0.05,
  MAX_COVG_QUANT = 0.95,
  FILTER_NA = TRUE,
  FILTER_SNP = TRUE,
  snp.list = NULL,
  FILTER_SOMATIC = TRUE,
  execute.lump = FALSE
)
}
\arguments{
\item{RNB_SET}{An object of type \code{\link[RnBeads]{RnBiseqSet-class}} for which analysis is to be performed.}

\item{WORK_DIR}{A path to a existing directory, in which the results are to be stored}

\item{analysis.name}{A string representing the dataset for which analysis is to be performed. Only used to create a folder with a 
descriptive name of the analysis.}

\item{SAMPLE_SELECTION_COL}{A column name in the phenotypic table of \code{RNB_SET} used to selected a subset of samples for
analysis that contain the string given in \code{SAMPLE_SELECTION_GREP}.}

\item{SAMPLE_SELECTION_GREP}{A string used for selecting samples in the column \code{SAMPLE_SELECTION_COL}.}

\item{REF_CT_COLUMN}{Column name in \code{RNB_SET} used to extract methylation information on the reference cell types.}

\item{PHENO_COLUMNS}{Vector of column names in the phenotypic table of \code{RNB_SET} that is kept and exported for further 
exploration.}

\item{PREPARE_TRUE_PROPORTIONS}{Flag indicating if true proportions are either available in \code{RNB_SET} or to be estimated 
with Houseman's reference-based deconvolution approach.}

\item{TRUE_A_TOKEN}{String present in the column names of \code{RNB_SET} used for selecting the true proportions of the corresponding
cell types.}

\item{HOUSEMAN_A_TOKEN}{Similar to \code{TRUE_A_TOKEN}, but not containing the true proportions, rather the estimated proportions
by Houseman's method.}

\item{ID_COLUMN}{Sample-specific ID column name in \code{RNB_SET}}

\item{FILTER_COVERAGE}{Flag indicating, if site-filtering based on coverage is to be conducted.}

\item{MIN_COVERAGE}{Minimum number of reads required in each sample for the site to be considered for adding to MeDeCom.}

\item{MIN_COVG_QUANT}{Lower quantile of coverages. Values lower than this value will be ignored for analysis.}

\item{MAX_COVG_QUANT}{Upper quantile of coverages. Values higher than this value will be ignored for analysis.}

\item{FILTER_NA}{Flag indicating if sites with any missing values are to be removed or not.}

\item{FILTER_SNP}{Flag indicating if annotated SNPs are to be removed from the list of sites according to RnBeads' SNP list.}

\item{snp.list}{Path to a file containing positions of known SNPs to be removed from the analysis, if \code{FILTER_SNP} is \code{TRUE}. The coordinates must be the 
provided in the same genome assembly as RNB_SET. The file must be a tab-separated value (tsv) file with only one header line an the following meaning of 
the rows: 1st row: chromosome, 2nd row: position of the SNP on the chromosome}

\item{FILTER_SOMATIC}{Flag indicating if only somatic probes are to be kept.}

\item{execute.lump}{Flag indicating if the LUMP algorithm is to be used for estimating the amount of immune cells in a particular sample.}
}
\value{
A list with four elements: \itemize{
          \item quality.filter The indices of the sites that survived quality filtering
}
}
\description{
This functions prepares sequencing data sets for a MeDeCom run.
}
