% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_analysis.R
\docType{data}
\name{rnb.set.example}
\alias{rnb.set.example}
\title{A small RnBeads object used to run the examples.}
\description{
A small RnBeads object used to run the examples.
}
\author{
Michael Scherer
}
