% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICA.R
\name{run.rnb.ICA}
\alias{run.rnb.ICA}
\title{run.rnb.ICA}
\usage{
run.rnb.ICA(rnb.set, conf.factor, ica.setting = NULL, nmin = 10,
  nmax = 30, ntry = 1, thr.sd = 0, alpha.fact = 1e-10,
  type = "remove", ncores = 1, out.folder = NULL, save.report = F,
  alpha.feat = 0.01)
}
\arguments{
\item{rnb.set}{An object of type \code{\link{RnBSet-class}} containing methylation information}

\item{conf.factor}{A vector of column names in the sample annotation sheet of \code{rnb.set} representing confounding factors to be 
removed.}

\item{ica.setting}{Optional argument. If specified a named vector of arguments to be used. Can be one of the following.}

\item{nmin}{Minimum number of components to be used}

\item{nmax}{Maximum number of components to be used}

\item{ntry}{Further argument for runICA}

\item{thr.sd}{Threshold for the standard deviation across samples. Only sites with a standard deviation larger than this
threshold are kept. 0 deactivates filtering.}

\item{alpha.fact}{Significance level for the factor}

\item{type}{Analysis type to be performed. Can be either \code{"remove"} to remove the effect of the confounding
factor or \code{"keep"} to export the sites that are linked to the confounding factor. Note that for
\code{"keep"}, the factor effect is not removed, but only reported.}

\item{out.folder}{Folder to store ICA's output and diagnostic plots}
}
\value{
The modified \code{rnb.set} object with updated methylation values. The effect of the confouding factor is removed
     using independent component analysis (ICA).
}
\description{
Performs ICA to remove the effect of a confounding factor from the methylation matrix
}
\author{
Michael Scherer. ICA code was generated by Peter Nazarov and Tony Kamoa
}
