% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_analysis.R
\name{start.analysis}
\alias{start.analysis}
\title{start.analysis}
\usage{
start.analysis(meth.data = NULL, rnb.set = NULL, cg_groups, Ks,
  work.dir, factorviz.outputs = FALSE, method = "MeDeCom", ...)
}
\arguments{
\item{meth.data}{A \code{matrix} or \code{data.frame} containing methylation information. If NULL, methylation information needs to be provided
through \code{rnb.set}}

\item{rnb.set}{An object of type \code{\link{RnBSet-class}} containing methylation and sample meta information.}

\item{cg_groups}{List of CpG indices used for the analysis. Can be computed by \code{\link{prepare_CG_subsets}}.}

\item{Ks}{Vector of integers used as components in MeDeCom.}

\item{work.dir}{Working directory for the analysis.}

\item{factorviz.outputs}{Flag indicating, if outputs should be stored to be compatible with FactorViz for data exploration}

\item{method}{The method to be used for deconvolution. Can be one of \code{MeDeCom}, \code{RefFreeCellMix} or \code{EDec}.}
}
\description{
Wrapper function to start one of the deconvolution algorithms \code{MeDeCom}, \code{RefFreeEWAS} or \code{EDec}
}
\author{
Michael Scherer
}
