% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_analysis.R
\name{start.decomp.pipeline}
\alias{start.decomp.pipeline}
\title{start.decomp.pipeline}
\usage{
start.decomp.pipeline(
  rnb.set,
  Ks,
  lambda.grid,
  work.dir = getwd(),
  factorviz.outputs = F,
  analysis.name = "Analysis",
  sample.selection.col = NA,
  sample.selection.grep = NA,
  pheno.cols = NA,
  id.column = rnb.getOption("identifiers.column"),
  normalization = "none",
  ref.ct.column = NA,
  ref.rnb.set = NULL,
  ref.rnb.ct.column = NA,
  prepare.true.proportions = F,
  true.A.token = NA,
  houseman.A.token = NA,
  estimate.houseman.prop = F,
  filter.beads = !is.null(rnb.set@covg.sites),
  min.n.beads = 3,
  filter.intensity = inherits(rnb.set, "RnBeadRawSet"),
  min.int.quant = 0.001,
  max.int.quant = 0.999,
  filter.na = TRUE,
  filter.context = TRUE,
  filter.cross.reactive = TRUE,
  execute.lump = FALSE,
  remove.ICA = FALSE,
  conf.fact.ICA = FALSE,
  ica.setting = NULL,
  filter.snp = TRUE,
  filter.sex.chromosomes = TRUE,
  snp.list = NULL,
  filter.coverage = hasCovg(rnb.set),
  min.coverage = 5,
  min.covg.quant = 0.05,
  max.covg.quant = 0.95,
  marker.selection = "var",
  n.markers = 5000,
  remove.correlated = FALSE,
  cor.threshold = "quantile",
  write.files = FALSE,
  n.prin.comp = 10,
  range.diff = 0.05,
  custom.marker.file = "",
  store.heatmaps = F,
  heatmap.sample.col = NULL,
  sample.subset = NULL,
  k.fixed = NULL,
  K.prior = NULL,
  opt.method = "MeDeCom.cppTAfact",
  startT = NULL,
  startA = NULL,
  folds = 10,
  cores = 1,
  itermax = 1000,
  ninit = 100,
  cluster.submit = FALSE,
  cluster.Rdir = NA,
  cluster.hostlist = "*",
  cluster.memlimit = "5G",
  cleanup = FALSE
)
}
\arguments{
\item{rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} for which analysis is to be performed.}

\item{Ks}{Vector of integers used as components in MeDeCom.}

\item{lambda.grid}{Vector of doubles representing the regularization parameter in MeDeCom.}

\item{work.dir}{A path to a existing directory, in which the results are to be stored}

\item{factorviz.outputs}{Flag indicating, if outputs should be stored to be compatible with FactorViz for data exploration}

\item{analysis.name}{A string representing the dataset for which analysis is to be performed. Only used to create a folder with a 
descriptive name of the analysis.}

\item{sample.selection.col}{A column name in the phenotypic table of \code{rnb.set} used to selected a subset of samples for
analysis that contain the string given in \code{sample.selection.col}.}

\item{sample.selection.grep}{A string used for selecting samples in the column \code{sample.selection.col}.}

\item{pheno.cols}{Vector of column names in the phenotypic table of \code{rnb.set} that is kept and exported for further 
exploration.}

\item{id.column}{Sample-specific ID column name in \code{rnb.set}}

\item{normalization}{normalization method to be performed before employing MeDeCom. Can be one of \code{"none","dasen","illumina","noob"} (BeadChip only).}

\item{ref.ct.column}{Column name in \code{rnb.set} used to extract methylation information on the reference cell types.}

\item{ref.rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} containing methylation information on reference cell types (BeadChip only).}

\item{ref.rnb.ct.column}{Column name in \code{ref.rnb.set} used to extract methylation information on the reference cell types (BeadChip only).}

\item{prepare.true.proportions}{Flag indicating if true proportions are either available in \code{rnb.set} or to be estimated 
with Houseman's reference-based deconvolution approach (BeadChip only).}

\item{true.A.token}{String present in the column names of \code{rnb.set} used for selecting the true proportions of the corresponding
cell types.}

\item{houseman.A.token}{Similar to \code{true.A.token}, but not containing the true proportions, rather the estimated proportions
by Houseman's method (BeadChip only).}

\item{estimate.houseman.prop}{If neither \code{true.A.token} nor \code{houseman.A.token} are given, the proportions of the reference
cell type are estimated with Houseman's approach (BeadChip only).}

\item{filter.beads}{Flag indicating, if site-filtering based on the number of beads available is to be conducted (BeadChip only).}

\item{min.n.beads}{Minimum number of beads required in each sample for the site to be considered for adding to MeDeCom (BeadChip only).}

\item{filter.intensity}{Flag indicating if sites should be removed according to the signal intensities (the lowest and highest quantiles
given by \code{min.int.quant} and \code{max.int.quant}) (BeadChip only).}

\item{min.int.quant}{Lower quantile of intensities which is to be removed (BeadChip only).}

\item{max.int.quant}{Upper quantile of intensities which is to be removed (BeadChip only).}

\item{filter.na}{Flag indicating if sites with any missing values are to be removed or not.}

\item{filter.context}{Flag indicating if only CG probes are to be kept (BeadChip only).}

\item{filter.cross.reactive}{Flag indicating if sites showing cross reactivity on the array are to be removed.}

\item{execute.lump}{Flag indicating if the LUMP algorithm is to be used for estimating the amount of immune cells in a particular sample.}

\item{remove.ICA}{Flag indicating if independent component analysis is to be executed to remove potential confounding factor.
If \code{TRUE},conf.fact.ICA needs to be specified.}

\item{conf.fact.ICA}{Column name in the sample annotation sheet representing a potential confounding factor.}

\item{ica.setting}{Optional argument setting up ICA.}

\item{filter.snp}{Flag indicating if annotated SNPs are to be removed from the list of sites according to RnBeads' SNP list.}

\item{filter.sex.chromosomes}{Flag indicating if only somatic probes are to be kept.
CPG FILTERING (BS)}

\item{snp.list}{Path to a file containing CpG IDs of known SNPs to be removed from the analysis, if \code{filter.snp} is \code{TRUE}.}

\item{filter.coverage}{Flag indicating, if site-filtering based on coverage is to be conducted (BS only).}

\item{min.coverage}{Minimum number of reads required in each sample for the site to be considered for adding to MeDeCom (BS only).}

\item{min.covg.quant}{Lower quantile of coverages. Values lower than this value will be ignored for analysis (BS only).}

\item{max.covg.quant}{Upper quantile of coverages. Values higher than this value will be ignored for analysis (BS only).
CG_SUBSET SELECTION}

\item{marker.selection}{A vector of strings representing marker selection methods. Available method are \itemize{
         \item{"\code{all}"} Using all sites available in the input.
         \item{"\code{pheno}"} Selected are the top \code{N_MARKERS} site that differ between the phenotypic
                groups defined in data preparation or by \code{\link{rnb.sample.groups}}. Those are
                selected by employing limma on the methylation matrix.
         \item{"\code{houseman2012}"} The 50k sites reported as cell-type specific in the Houseman's reference-
                based deconvolution. See Houseman et.al. 2012.
         \item{"\code{houseman2014}"} Selects the sites said to be linked to cell type composition by \code{RefFreeEWAS},
                which is similar to surrogate variable analysis. See Houseman et.al. 2014.
         \item{"\code{jaffe2014}"} The sites stated as related to cell-type composition Jaffe et.al. 2014.
         \item{"\code{rowFstat}"} Markers are selected as those found to be associated to the reference cell
                types with F-statistics. If this option is selected, \code{REF_DATA_SET} and \code{REF_PHENO_COLUMN}
                need to be specified.
         \item{"\code{random}"} Sites are randomly selected.
         \item{"\code{pca}"} Sites are selected as those with most influence on the principal components.
         \item{"\code{var}"} Selects the most variable sites.
         \item{"\code{hybrid}"} Selects (N_MARKERS/2) most variable and (N_MARKERS/2) random sites.
         \item{"\code{range}"} Selects the sites with the largest difference between minimum and maximum
              across samples.
         \item{"\code{pcadapt}"} Uses principal component analysis as implemented in the \code{"bigstats"}
              R package to determine sites that are significantly linked to the potential cell types. This
              requires specifying K a priori (argument \code{K.prior}). We thank Florian Prive and Sophie
              Achard for providing the idea and parts of the codes.
         \item{"\code{edec_stage0}} Employs EDec's stage 0 to infer cell-type specific markers. By default
              EDec's example reference data is provided. If a specific data set is to be provided, it needs
              to be done through \code{REF_DATA_SET}.
         \item{"\code{custom}"} Specifying a custom file with indices.
}}

\item{n.markers}{The number of sites to be selected. Defaults to 5000.}

\item{remove.correlated}{Flag indicating if highly correlated features are to be removed.}

\item{cor.threshold}{Numeric indicating a threshold above which sites are not to be considered in the feature selection.
If \code{"quantile"}, sites correlated higher than the 95th quantile are removed.}

\item{write.files}{Flag indicating if the selected sites are to be stored on disk.}

\item{n.prin.comp}{Optional argument deteriming the number of prinicipal components used for selecting the most important sites.}

\item{range.diff}{Optional argument specifying the difference between maximum and minimum required.}

\item{custom.marker.file}{Optional argument containing a file that specifies the indices used for employing MeDeCom.}

\item{store.heatmaps}{Flag indicating if a heatmap of the selected input sites is to be create from the input methylation matrix.
The files are then stored in the 'heatmaps' folder in WD.}

\item{heatmap.sample.col}{Column name in the phenotypic table of \code{rnb.set}, used for creating a color scheme in the heatmap.}

\item{sample.subset}{Vector of indices of samples to be included in the analysis. If \code{NULL}, all samples are included.}

\item{k.fixed}{Columns in the T matrix that should be fixed. If \code{NULL}, no columns are fixed.}

\item{K.prior}{K determined from visual inspection. Only has an influence, if \code{MARKER_SELECTION="pcadapt"}.}

\item{opt.method}{Optimization method to be used. Either MeDeCom.quadPen or MeDeCom.cppTAfact (default).}

\item{startT}{Inital matrix for T.}

\item{startA}{Initial matrix for A.}

\item{folds}{Integer representing the number of folds used in the analysis.}

\item{cores}{Integer representing the number of cores to be used in the analysis.}

\item{itermax}{Maximum number of iterations}

\item{ninit}{Number if initialtions.}

\item{cluster.submit}{Flag indicating, if the jobs are to be submitted to a scientific compute cluster (only SGE supported).}

\item{cluster.Rdir}{Path to an executable version of R.}

\item{cluster.hostlist}{Regular expression, on which basis hosts are selected in the cluster environment.}

\item{cluster.memlimit}{the \code{memlimit} resource value of the cluster submission.}

\item{cleanup}{Flag indicating if temprary files are to be deleted.}
}
\value{
An object of type \code{\link{MeDeComSet}} containing the results of the MeDeCom experiment.
}
\description{
Main workhorse of the DecompPipeline R-package. Performs preprocessing (\code{\link{prepare.data}} or \code{\link{prepare.data.BS}}),
CpG subset selection (\code{\link{prepare.CG.subsets}}) and deconvolution (\code{\link{start.medecom.analysis}}, \code{\link{start.refreeewas.analysis}}, \code{\link{start.edec.analysis}})
}
\author{
Michael Scherer
}
