% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_analysis.R
\name{start.medecom.analysis}
\alias{start.medecom.analysis}
\title{start.medecom.analysis}
\usage{
start.medecom.analysis(
  meth.data = NULL,
  rnb.set = NULL,
  work.dir = getwd(),
  cg.groups,
  Ks,
  lambda.grid,
  sample.subset = NULL,
  k.fixed = NULL,
  write.files = TRUE,
  factorviz.outputs = F,
  opt.method = "MeDeCom.cppTAfact",
  startT = NULL,
  startA = NULL,
  trueT = NULL,
  trueA = NULL,
  analysis.name = "MeDeComRun",
  folds = 10,
  cores = 1,
  itermax = 1000,
  ninit = 100,
  cluster.submit = FALSE,
  cluster.architecture = "SGE",
  cluster.R.dir = NA,
  cluster.hostlist = "*",
  cluster.memlimit = "5G",
  n.cpus = 1,
  cleanup = FALSE,
  analysis.info = NULL,
  lambda.grid.type = "standard",
  analysis.token = "customAnalysis"
)
}
\arguments{
\item{meth.data}{A \code{matrix} or \code{data.frame} containing methylation information. If NULL, methylation information needs to be provided
through \code{rnb.set}}

\item{rnb.set}{An object of type \code{\link{RnBSet-class}} containing methylation and sample meta information.}

\item{work.dir}{Working directory for the analysis.}

\item{cg.groups}{List of CpG indices used for the analysis. Can be computed by \code{\link{prepare.CG.subsets}}.}

\item{Ks}{Vector of integers used as components in MeDeCom.}

\item{lambda.grid}{Vector of doubles representing the regularization parameter in MeDeCom.}

\item{sample.subset}{Vector of indices of samples to be included in the analysis. If \code{NULL}, all samples are included.}

\item{k.fixed}{Columns in the T matrix that should be fixed. If \code{NULL}, no columns are fixed.}

\item{write.files}{Flag indicating if intermediate results are to be stored.}

\item{factorviz.outputs}{Flag indicating, if outputs should be stored to be compatible with FactorViz for data exploration}

\item{opt.method}{Optimization method to be used. Either MeDeCom.quadPen or MeDeCom.cppTAfact (default).}

\item{startT}{Inital matrix for T.}

\item{startA}{Initial matrix for A.}

\item{trueT}{True value for the T matrix.}

\item{trueA}{True value for the A matrix.}

\item{analysis.name}{Name of the analysis.}

\item{folds}{Integer representing the number of folds used in the analysis.}

\item{cores}{Integer representing the number of cores to be used in the analysis.}

\item{itermax}{Maximum number of iterations}

\item{ninit}{Number if initialtions.}

\item{cluster.submit}{Flag indicating, if the jobs are to be submitted to a scientific compute cluster (only SGE and SLURM supported).}

\item{cluster.architecture}{The cluster architecture used. Currently \code{'SGE'} and \code{'SLURM'} supported}

\item{cluster.R.dir}{Path to an executable version of R.}

\item{cluster.hostlist}{Regular expression, on which basis hosts are selected in the cluster environment.}

\item{cluster.memlimit}{the \code{memlimit} resource value of the cluster submission.}

\item{n.cpus}{Number of CPUs used per node (SLURM only)}

\item{cleanup}{Flag indicating if temprary files are to be deleted.}

\item{analysis.info}{Information to be saved about the analysis. Just stored as info.}

\item{lambda.grid.type}{String represent the lambda grid that was chosen. Just stored as info.}

\item{analysis.token}{String specifying the type of analysis that was conducted. Just stored as info.}
}
\value{
An object of type \code{\link{MeDeComSet}} containing the results of the MeDeCom experiment.
}
\description{
Wrapper for runMeDeCom, for data preprocessed through the DecompPipeline
}
\author{
Pavlo Lutsik, Michael Scherer
}
