% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{filter.annotation}
\alias{filter.annotation}
\title{filter.annotation}
\usage{
filter.annotation(rnb.set, snp = TRUE, snp.list, somatic = TRUE,
  context = TRUE, cross.reactive = T, qual.filter = NULL)
}
\arguments{
\item{rnb.set}{An object of type \code{\link[RnBeads]{RnBSet-class}} containg annotation information.}

\item{snp}{Flag indicating if snps are to be removed. Either SNPs annotated in the RnBeads annotation object of specified as an
additional file with one SNP identifier per row in \code{snp.list}.}

\item{snp.list}{Path to a file containing known SNPs. One SNP identifier should be present per row.}

\item{somatic}{Flag indicating if only somatic probes are to be kept.}

\item{context}{Flag indicating if probes in non-CpG context are to be removed.}

\item{cross.reactive}{Flag indicating if cross-reactive probes are to be removed.}

\item{qual.filter}{Vector of indices removed during quality filtering.}
}
\value{
A vector of indices of sites surviving the annotation filter criteria.
}
\description{
This function removes sites in SNP, sex-chromosomal, or non-CG context.
}
