% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpG_subsets.R
\name{prepare_CG_subsets}
\alias{prepare_CG_subsets}
\title{prepare_CG_subsets}
\usage{
prepare_CG_subsets(rnb.set, MARKER_SELECTION, N_MARKERS = 5000,
  WRITE_FILES = FALSE, WD = NA, REF_DATA_SET = NULL,
  REF_PHENO_COLUMN = NULL, N_PRIN_COMP = 10, RANGE_DIFF = 0.05,
  CUSTOM_MARKER_FILE = "", store.heatmaps = F, heatmap.sample.col = NULL)
}
\arguments{
\item{rnb.set}{An object of type \code{\link{RnBSet-class}} containing methylation, sample and optional coverage information.}

\item{MARKER_SELECTION}{A vector of strings representing marker selection methods. Available method are \itemize{
         \item{"\code{pheno}"} Selected are the top \code{N_MARKERS} site that differ between the phenotypic
                groups defined in data preparation or by \code{\link{rnb.sample.groups}}. Those are
                selected by employing limma on the methylation matrix.
         \item{"\code{houseman2012}"} The 50k sites reported as cell-type specific in the Houseman's reference-
                based deconvolution. See Houseman et.al. 2012.
         \item{"\code{houseman2014}"} Selects the sites said to be linked to cell type composition by \code{RefFreeEWAS},
                which is similar to surrogate variable analysis. See Houseman et.al. 2014.
         \item{"\code{jaffe2014}"} The sites stated as related to cell-type composition Jaffe et.al. 2014.
         \item{"\code{rowFstat}"} Markers are selected as those found to be associated to the reference cell
                types with F-statistics. If this option is selected, \code{REF_DATA_SET} and \code{REF_PHENO_COLUMN}
                need to be specified.
         \item{"\code{random}"} Sites are randomly selected.
         \item{"\code{pca}"} Sites are selected as those with most influence on the principal components.
         \item{"\code{var}"} Selects the most variable sites.
         \item{"\code{hybrid}"} Selects (N_MARKERS/2) most variable and (N_MARKERS/2) random sites.
         \item{"\code{range}"} Selects the sites with the largest difference between minimum and maximum
              across samples.
         \item{"\code{custom}"} Specifying a custom file with indices.
}}

\item{N_MARKERS}{The number of sites to be selected. Defaults to 5000.}

\item{WRITE_FILES}{Flag indicating if the selected sites are to be stored on disk.}

\item{WD}{Path to the working directory used for analyis, or data preparation.}

\item{REF_DATA_SET}{An object of type \code{\link{RnBSet-class}} or a path to such an object stored on disk, 
if \code{rowFstat} is selected.}

\item{REF_PHENO_COLUMN}{Optional argument stating the column name of the phenotypic table of \code{REF_DATA_SET} with
the reference cell type.}

\item{N_PRIN_COMP}{Optional argument deteriming the number of prinicipal components used for selecting the most important sites.}

\item{RANGE_DIFF}{Optional argument specifying the difference between maximum and minimum required.}

\item{CUSTOM_MARKER_FILE}{Optional argument containing a file that specifies the indices used for employing MeDeCom.}

\item{store.heatmaps}{Flag indicating if a heatmap of the selected input sites is to be create from the input methylation matrix.
The files are then stored in the 'heatmaps' folder in WD.}

\item{heatmap.sample.col}{Column name in the phenotypic table of \code{rnb.set}, used for creating a color scheme in the heatmap.}
}
\value{
List of indices, one entry for each marker selection method specified by \code{MARKER_SELECTION}. The indices correspond
         to the sites that should be used in \code{rnb.set}.
}
\description{
This routine selects a subset of CpGs sites used for MeDeCom analysis. Different selection methods are supported.
}
\details{
For methods "\code{houseman2012}" and "\code{jaffe2014}", a predefined set of markers is used. Since those correspond to
         absolute indices on the chip, the provided \code{rnb.set} must not be preprocessed and therefore still contain all sites.
         For the other metods, you may used \code{\link{prepare_data}} to filter sites for quality and context.
}
\references{
\itemize{
            \item{1.} Houseman, E. A., Accomando, W. P., Koestler, D. C., Christensen, B. C., Marsit, C. J., Nelson, H. H., ..., Kelsey, K.
                T. (2012). DNA methylation arrays as surrogate measures of cell mixture distribution. BMC Bioinformatics, 13. 
            \item{2.} Houseman, E. A., Molitor, J., & Marsit, C. J. (2014). Reference-free cell mixture adjustments in analysis of 
                DNA methylation data. Bioinformatics, 30(10), 1431-1439. https://doi.org/10.1093/bioinformatics/btu029
            \item{3.} Jaffe, A. E., & Irizarry, R. A. (2014). Accounting for cellular heterogeneity is critical in epigenome-wide 
                association studies. Genome Biology, 15(2), R31. https://doi.org/10.1186/gb-2014-15-2-r31
}
}
